break index into more components
